extract decorateUserRes to module (#186)
